Add missing static_cast
git-svn-id: https://svn.dealii.org/trunk@8739 0785d39b-7218-0410-832d-ea1e28bc413d

+++ tests/deal.II/subdomain_ids.cc
@@ -165,7 +165,8 @@ void test ()
 		  ExcNumberMismatch(std::count (selected_dofs.begin(),
 						selected_dofs.end(),
 						true),
-				    std::pow(static_cast<double>(cells_per_direction/2+1),dim)));
+				    static_cast<unsigned int>(std::pow(static_cast<double>(
+				      cells_per_direction/2+1),dim))));
 	}
       deallog << "Check 4 (dim=" << dim << ") ok" << std::endl;      
     };
